# apps/core-ledger-api/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Core Ledger API",
    "description": "Financial Ledger REST API for institutional financial clients",
    "contact": {
      "name": "Core Ledger Team",
      "email": "support@coreledger.com"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:7109",
      "description": "Local development (HTTPS)"
    },
    {
      "url": "http://localhost:5071",
      "description": "Local development (HTTP)"
    }
  ],
  "paths": {
...
    "/api/transactions/types/{id}": {
      "get": {
        "tags": [
          "Transaction Types"
        ],
        "operationId": "GetTransactionTypeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.DTOs.TransactionTypeDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "operationId": "GetAllTransactions",
        "parameters": [
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.Models.PagedResult`1[[CoreLedger.Application.DTOs.TransactionDto, CoreLedger.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Transactions"
        ],
        "operationId": "CreateTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoreLedger.Application.DTOs.CreateTransactionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.DTOs.TransactionDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/api/transactions/{id}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "operationId": "GetTransactionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.DTOs.TransactionDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Transactions"
        ],
        "operationId": "UpdateTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoreLedger.Application.DTOs.UpdateTransactionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "GetCurrentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.DTOs.UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoreLedger.Application.DTOs.CreateTransactionDto": {
        "type": "object",
        "properties": {
          "fundId": {
            "type": "integer",
            "format": "int32"
          },
          "securityId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "transactionSubTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "tradeDate": {
            "type": "string",
            "format": "date-time"
          },
          "settleDate": {
            "type": "string",
            "format": "date-time"
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "idempotencyKey": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CoreLedger.Application.DTOs.TransactionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fundId": {
            "type": "integer",
            "format": "int32"
          },
          "fundCode": {
            "type": "string",
            "nullable": true
          },
          "fundName": {
            "type": "string",
            "nullable": true
          },
          "securityId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "securityTicker": {
            "type": "string",
            "nullable": true
          },
          "securityName": {
            "type": "string",
            "nullable": true
          },
          "transactionSubTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "transactionSubTypeDescription": {
            "type": "string",
            "nullable": true
          },
          "transactionTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "transactionTypeDescription": {
            "type": "string",
            "nullable": true
          },
          "tradeDate": {
            "type": "string",
            "format": "date-time"
          },
          "settleDate": {
            "type": "string",
            "format": "date-time"
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "statusId": {
            "type": "integer",
            "format": "int32"
          },
          "statusDescription": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CoreLedger.Application.DTOs.TransactionTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "shortDescription": {
            "type": "string",
            "nullable": true
          },
          "longDescription": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CoreLedger.Application.DTOs.UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "fundId": {
            "type": "integer",
            "format": "int32"
          },
          "securityId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "transactionSubTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "tradeDate": {
            "type": "string",
            "format": "date-time"
          },
          "settleDate": {
            "type": "string",
            "format": "date-time"
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "statusId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CoreLedger.Application.DTOs.UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "authProviderId": {
            "type": "string",
            "nullable": true
          },
          "provider": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below.\n\nExample: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}